+++ src/tests/reducers/usersReducer.test.js
@@ -1,5 +1,5 @@
 import usersReducer from '../../reducers/usersReducer'
-import { GET_USERS } from '../../types'
+import { GET_USERS, POST_SIGNUP_INFO } from '../../types'
 
 describe('reduces users', () => {
   it('defaults to empty projects if none are passed in', () => {
@@ -14,4 +14,10 @@ describe('reduces users', () => {
       })
     ).toEqual(['User to be added to store'])
   })
+  expect(
+    usersReducer([], {
+      type: POST_SIGNUP_INFO,
+      payload: ['User to be added to store']
+    })
+  ).toEqual(['User to be added to store'])
 })
