+++ src/tests/store/store.test.js
@@ -11,4 +11,15 @@ describe('Store', () => {
       users: ['Run the tests']
     })
   })
+
+  it('should update after signup', async () => {
+    await store.dispatch({
+      type: 'POST_SIGNUP_INFO',
+      payload: ['Rerun them']
+    })
+
+    expect(store.getState()).toEqual({
+      users: ['Rerun them']
+    })
+  })
 })
